Refactor namespace
