+++ src/Extension/ApiPlatform/Doctrine/Orm/KundeOwnedByCurrentVermittlerUserExtension.php
@@ -2,17 +2,26 @@
 
 declare(strict_types=1);
 
-namespace App\Extension\Doctrine\Doctrine\Orm;
+namespace App\Extension\ApiPlatform\Doctrine\Orm;
 
 use ApiPlatform\Doctrine\Orm\Extension\QueryCollectionExtensionInterface;
 use ApiPlatform\Doctrine\Orm\Extension\QueryItemExtensionInterface;
 use ApiPlatform\Doctrine\Orm\Util\QueryNameGeneratorInterface;
 use ApiPlatform\Metadata\Operation;
 use App\Entity\Kunde;
+use App\Entity\VermittlerUser;
+use Doctrine\ORM\Query\Expr\Join;
 use Doctrine\ORM\QueryBuilder;
+use Symfony\Bundle\SecurityBundle\Security;
 
-class KundeNotGeloeschtExtension implements QueryCollectionExtensionInterface, QueryItemExtensionInterface
+class KundeOwnedByCurrentVermittlerUserExtension
+    implements QueryCollectionExtensionInterface, QueryItemExtensionInterface
 {
+    public function __construct(
+        private Security $security
+    ) {
+    }
+
     public function applyToCollection(
         QueryBuilder $queryBuilder,
         QueryNameGeneratorInterface $queryNameGenerator,
@@ -40,10 +49,18 @@ class KundeNotGeloeschtExtension implements QueryCollectionExtensionInterface, Q
             return;
         }
 
+        $user = $this->security->getUser();
+        if ($user === null) {
+            return;
+        }
+
         $rootAlias = $queryBuilder->getRootAliases()[0];
-        $queryBuilder->andWhere(
-            sprintf('%s.geloescht = :zero', $rootAlias)
+        $queryBuilder->join(
+            join: VermittlerUser::class,
+            alias: 'user',
+            conditionType: Join::WITH,
+            condition: sprintf('user = :current_user AND user.vermittler = %s.vermittler', $rootAlias)
         );
-        $queryBuilder->setParameter('zero', 0);
+        $queryBuilder->setParameter('current_user', $user);
     }
 }
